requirements.txt
Flask  
datetime 
uuid
Flask-SQLAlchemy
PyJWT

